chore: update notification reddot style
