+++ frontend/appflowy_flutter/lib/shared/red_dot.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/widgets.dart';
+
+class NotificationRedDot extends StatelessWidget {
+  const NotificationRedDot({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      width: 6,
+      height: 6,
+      clipBehavior: Clip.antiAlias,
+      decoration: ShapeDecoration(
+        color: const Color(0xFFFF2214),
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(20),
+        ),
+      ),
+    );
+  }
+}
