+++ frontend/appflowy_flutter/lib/workspace/presentation/notifications/widgets/notification_button.dart
@@ -1,5 +1,6 @@
 import 'package:appflowy/generated/flowy_svgs.g.dart';
 import 'package:appflowy/generated/locale_keys.g.dart';
+import 'package:appflowy/shared/red_dot.dart';
 import 'package:appflowy/startup/startup.dart';
 import 'package:appflowy/user/application/reminder/reminder_bloc.dart';
 import 'package:appflowy/workspace/application/menu/sidebar_sections_bloc.dart';
@@ -7,7 +8,6 @@ import 'package:appflowy/workspace/application/settings/notifications/notificati
 import 'package:appflowy/workspace/presentation/notifications/notification_dialog.dart';
 import 'package:appflowy_popover/appflowy_popover.dart';
 import 'package:easy_localization/easy_localization.dart';
-import 'package:flowy_infra/theme_extension.dart';
 import 'package:flowy_infra_ui/flowy_infra_ui.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -43,33 +43,35 @@ class _NotificationButtonState extends State<NotificationButton> {
       child: BlocBuilder<NotificationSettingsCubit, NotificationSettingsState>(
         builder: (notificationSettingsContext, notificationSettingsState) {
           return BlocBuilder<ReminderBloc, ReminderState>(
-            builder: (context, state) => notificationSettingsState
-                    .isShowNotificationsIconEnabled
-                ? FlowyTooltip(
-                    message: LocaleKeys.notificationHub_title.tr(),
-                    child: AppFlowyPopover(
-                      mutex: mutex,
-                      direction: PopoverDirection.bottomWithLeftAligned,
-                      constraints:
-                          const BoxConstraints(maxHeight: 500, maxWidth: 425),
-                      windowPadding: EdgeInsets.zero,
-                      margin: EdgeInsets.zero,
-                      popupBuilder: (_) =>
-                          NotificationDialog(views: views, mutex: mutex),
-                      child: SizedBox.square(
-                        dimension: 24.0,
-                        child: FlowyButton(
-                          useIntrinsicWidth: true,
-                          margin: EdgeInsets.zero,
-                          text: _buildNotificationIcon(
-                            context,
-                            state.hasUnreads,
+            builder: (context, state) {
+              final hasUnreads = state.reminders.any((r) => !r.isRead);
+              return notificationSettingsState.isShowNotificationsIconEnabled
+                  ? FlowyTooltip(
+                      message: LocaleKeys.notificationHub_title.tr(),
+                      child: AppFlowyPopover(
+                        mutex: mutex,
+                        direction: PopoverDirection.bottomWithLeftAligned,
+                        constraints:
+                            const BoxConstraints(maxHeight: 500, maxWidth: 425),
+                        windowPadding: EdgeInsets.zero,
+                        margin: EdgeInsets.zero,
+                        popupBuilder: (_) =>
+                            NotificationDialog(views: views, mutex: mutex),
+                        child: SizedBox.square(
+                          dimension: 24.0,
+                          child: FlowyButton(
+                            useIntrinsicWidth: true,
+                            margin: EdgeInsets.zero,
+                            text: _buildNotificationIcon(
+                              context,
+                              hasUnreads,
+                            ),
                           ),
                         ),
                       ),
-                    ),
-                  )
-                : const SizedBox.shrink(),
+                    )
+                  : const SizedBox.shrink();
+            },
           );
         },
       ),
@@ -86,16 +88,10 @@ class _NotificationButtonState extends State<NotificationButton> {
           ),
         ),
         if (hasUnreads)
-          Positioned(
-            bottom: 2,
-            right: 2,
-            child: DecoratedBox(
-              decoration: BoxDecoration(
-                shape: BoxShape.circle,
-                color: AFThemeExtension.of(context).warning,
-              ),
-              child: const SizedBox(height: 8, width: 8),
-            ),
+          const Positioned(
+            top: 3,
+            right: 5,
+            child: NotificationRedDot(),
           ),
       ],
     );
